Check if an argument has been given for the application and exit
Without any arguments, the program will just crash. A command-line
option parser with multiple flag options would be cool. Then it could
just default to the help menu instead of crashing.

But, until I get around to that or forget about it: a simple warning to
let the user know what went wrong and a proper exit code should suffice?